backend/requirements.txt
flask
psycopg2-binary
pytest
flake8
